pyproject.toml
[project]
name = "polymarket-bot"
version = "0.1.0"
description = "Arbitrage bot for Polymarket 15-minute crypto markets"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "py-clob-client>=0.34.0",
    "websockets>=12.0",
    "redis>=5.0.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.9.0",
    "pandas>=2.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
